requirements.txt
click==7.1.1
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
nltk==3.4.5
Pillow==7.1.1
pytesseract==0.3.3
six==1.14.0
Werkzeug==1.0.1
